requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.1
attrs==21.2.0
bcrypt==3.2.0
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
cryptography==36.0.0
frozenlist==1.2.0
idna==3.3
lastpass-python==0.3.2
multidict==5.2.0
multivolumefile==0.2.3
paramiko==2.8.1
py7zr==0.16.2
pybcj==0.5.0
pycparser==2.21
pycryptodome==3.11.0
pycryptodomex==3.11.0
PyNaCl==1.4.0
pyotp==2.6.0
pyppmd==0.17.3
pysftp==0.2.9
pyzstd==0.15.0
requests==2.26.0
six==1.16.0
slackclient==2.9.3
texttable==1.6.4
typing-extensions==4.0.0
urllib3==1.26.7
yarl==1.7.2
